List jOOQ in the database access section
Hi there,

I was made aware of this list by a jOOQ user and was wondering if you're also accepting Java APIs in your list. We have many Scala/jOOQ users and also a couple of paying customers using this combination with a lot of joy. Of course, the usual interoperability issues (boxed number types, `Option`) appear also here, but in general, writing SQL in Scala with jOOQ is awesome!

If it doesn't match, feel free to reject.

Cheers
Lukas