Updated readme file
Added a really cool library that helped me out a lot today. Unlike standard swagger libraries which pollute the code with ugly annotations, this one allows the user to only modify the routes file in order to generate the docs. It's also very easy to redirect an endpoint directly to Swagger UI so that the generated JSON/YAML file is shown in the Swagger interface.